+++ src/main/java/database/ParFileUser.java
@@ -1,20 +1,19 @@
-package tutorial;
+package database;
 
 import java.sql.*;
 import java.util.ArrayList;
 
-public class OrgDsFileUser implements OrgDsGateway {
+public class ParFileUser {
     public static void main(String[] args) {
-        OrgDsFileUser a = new OrgDsFileUser();
-        System.out.println(a.organizerSearch("s"));
     }
-    public void utilStoreOrg(String username, String password){
+
+    public void utilStorePar(String username, String password){
         Statement stmt = null;
         Connection conn = null;
         try {
             Class.forName("com.mysql.jdbc.Driver");
             conn = DriverManager.getConnection("jdbc:mysql://localhost:3306/db2", "root", "1234");
-            String sql = "insert into orgfile(username, password) values('" + username + "','" + password + "');" ;
+            String sql = "insert into parfile(username, password) values('" + username + "','" + password + "');" ;
             stmt = conn.createStatement();
             int count = stmt.executeUpdate(sql);
             System.out.println(sql);
@@ -49,13 +48,13 @@ public class OrgDsFileUser implements OrgDsGateway {
 
     }
 
-    public void utilDeleteOrg(String username){
+    public void utilDeletePar(String username){
         Statement stmt = null;
         Connection conn = null;
         try {
             Class.forName("com.mysql.jdbc.Driver");
             conn = DriverManager.getConnection("jdbc:mysql://localhost:3306/db2", "root", "1234");
-            String sql = "delete from orgfile where username = '" + username + "';";
+            String sql = "delete from parfile where username = '" + username + "';";
             stmt = conn.createStatement();
             int count = stmt.executeUpdate(sql);
             System.out.println(sql);
@@ -89,13 +88,13 @@ public class OrgDsFileUser implements OrgDsGateway {
         }
     }
 
-    public void utilAddOrgPastEvent(String org_username, String event_title){
+    public void utiladdparfollowing(String par_username, String org_username){
         Statement stmt = null;
         Connection conn = null;
         try {
             Class.forName("com.mysql.jdbc.Driver");
             conn = DriverManager.getConnection("jdbc:mysql://localhost:3306/db2", "root", "1234");
-            String sql = "insert into past_events_for_org(org_username, event_title) values('" + org_username + "','" + event_title + "');" ;
+            String sql = "insert into follow_org_par(par_username, org_username) values('" + par_username + "','" + org_username + "');" ;
             stmt = conn.createStatement();
             int count = stmt.executeUpdate(sql);
             System.out.println(sql);
@@ -131,13 +130,178 @@ public class OrgDsFileUser implements OrgDsGateway {
 
     }
 
-    public void utilDeleteOrgPastEvent(String org_username, String event_title) {
+    public void utilDeleteParFollowOrg(String par_username, String org_username) {
+            Statement stmt = null;
+            Connection conn = null;
+            try {
+                Class.forName("com.mysql.jdbc.Driver");
+                conn = DriverManager.getConnection("jdbc:mysql://localhost:3306/db2", "root", "1234");
+                String sql = "delete from follow_org_par where par_username = '" + par_username + "' and org_username = '" + org_username + "';";
+                stmt = conn.createStatement();
+                int count = stmt.executeUpdate(sql);
+                System.out.println(sql);
+                if (count > 0) {
+                    System.out.println("Success");
+                } else {
+                    System.out.println("Failure");
+                }
+
+            } catch (ClassNotFoundException e) {
+                e.printStackTrace();
+            } catch (SQLException e) {
+                e.printStackTrace();
+            } finally {
+                if (stmt != null) {
+                    try {
+                        stmt.close();
+                    } catch (SQLException e) {
+                        e.printStackTrace();
+                    }
+                }
+                if (conn != null) {
+                    try {
+                        conn.close();
+                    } catch (SQLException e) {
+                        e.printStackTrace();
+                    }
+                }
+
+            }
+
+        }
+
+    public void utilAddParPastevent(String par_username, String event_title){
         Statement stmt = null;
         Connection conn = null;
         try {
             Class.forName("com.mysql.jdbc.Driver");
             conn = DriverManager.getConnection("jdbc:mysql://localhost:3306/db2", "root", "1234");
-            String sql = "delete from follow_org_par where org_username = '" + org_username + "' and event_title = '" + event_title + "';";
+            String sql = "insert into past_events_for_par(par_username, event_title) values('" + par_username + "','" + event_title + "');" ;
+            stmt = conn.createStatement();
+            int count = stmt.executeUpdate(sql);
+            System.out.println(sql);
+            if (count > 0 ){
+                System.out.println("Success");
+            }
+            else {
+                System.out.println("Failure");
+            }
+
+        } catch (ClassNotFoundException e) {
+            e.printStackTrace();
+        } catch (SQLException e) {
+            e.printStackTrace();
+        }finally {
+            if (stmt != null){
+                try {
+                    stmt.close();
+                } catch (SQLException e) {
+                    e.printStackTrace();
+                }
+            }
+            if (conn != null){
+                try {
+                    conn.close();
+                } catch (SQLException e) {
+                    e.printStackTrace();
+                }
+            }
+
+        }
+
+
+    }
+    public void utilDeleteParPastevent(String par_username, String event_title){
+        Statement stmt = null;
+        Connection conn = null;
+        try {
+            Class.forName("com.mysql.jdbc.Driver");
+            conn = DriverManager.getConnection("jdbc:mysql://localhost:3306/db2", "root", "1234");
+            String sql = "delete from past_events_for_par where par_username = '" + par_username + "' and event_title = '" + event_title + "';";
+            stmt = conn.createStatement();
+            int count = stmt.executeUpdate(sql);
+            System.out.println(sql);
+            if (count > 0) {
+                System.out.println("Success");
+            } else {
+                System.out.println("Failure");
+            }
+
+        } catch (ClassNotFoundException e) {
+            e.printStackTrace();
+        } catch (SQLException e) {
+            e.printStackTrace();
+        } finally {
+            if (stmt != null) {
+                try {
+                    stmt.close();
+                } catch (SQLException e) {
+                    e.printStackTrace();
+                }
+            }
+            if (conn != null) {
+                try {
+                    conn.close();
+                } catch (SQLException e) {
+                    e.printStackTrace();
+                }
+            }
+
+        }
+
+
+
+    }
+
+    public void utilAddParUpcomingEvent(String par_username, String event_title){
+        Statement stmt = null;
+        Connection conn = null;
+        try {
+            Class.forName("com.mysql.jdbc.Driver");
+            conn = DriverManager.getConnection("jdbc:mysql://localhost:3306/db2", "root", "1234");
+            String sql = "insert into upcoming_events_for_par(par_username, event_title) values('" + par_username + "','" + event_title + "');" ;
+            stmt = conn.createStatement();
+            int count = stmt.executeUpdate(sql);
+            System.out.println(sql);
+            if (count > 0 ){
+                System.out.println("Success");
+            }
+            else {
+                System.out.println("Failure");
+            }
+
+        } catch (ClassNotFoundException e) {
+            e.printStackTrace();
+        } catch (SQLException e) {
+            e.printStackTrace();
+        }finally {
+            if (stmt != null){
+                try {
+                    stmt.close();
+                } catch (SQLException e) {
+                    e.printStackTrace();
+                }
+            }
+            if (conn != null){
+                try {
+                    conn.close();
+                } catch (SQLException e) {
+                    e.printStackTrace();
+                }
+            }
+
+        }
+
+
+    }
+
+    public void utilDeleteParUpcomingevent(String par_username, String event_title){
+        Statement stmt = null;
+        Connection conn = null;
+        try {
+            Class.forName("com.mysql.jdbc.Driver");
+            conn = DriverManager.getConnection("jdbc:mysql://localhost:3306/db2", "root", "1234");
+            String sql = "delete from upcoming_events_for_par where par_username = '" + par_username + "' and event_title = '" + event_title + "';";
             stmt = conn.createStatement();
             int count = stmt.executeUpdate(sql);
             System.out.println(sql);
@@ -171,28 +335,34 @@ public class OrgDsFileUser implements OrgDsGateway {
 
     }
 
-    public void utilAddOrgUnpublishedEvent(String org_username, String event_title){
+    public ArrayList<String> utilGetAllFollowing(String par_username){
         Statement stmt = null;
         Connection conn = null;
+        ResultSet rs = null;
+        ArrayList l = new ArrayList<String>(0);
         try {
             Class.forName("com.mysql.jdbc.Driver");
             conn = DriverManager.getConnection("jdbc:mysql://localhost:3306/db2", "root", "1234");
-            String sql = "insert into unpublished_events_for_org(org_username, event_title) values('" + org_username + "','" + event_title + "');" ;
+            String sql = "select org_username from follow_org_par where par_username = '" + par_username + "';";
             stmt = conn.createStatement();
-            int count = stmt.executeUpdate(sql);
+            rs = stmt.executeQuery(sql);
             System.out.println(sql);
-            if (count > 0 ){
-                System.out.println("Success");
+//            rs.next();
+            while (rs.next()){
+                l.add(rs.getString(1));
             }
-            else {
-                System.out.println("Failure");
-            }
-
         } catch (ClassNotFoundException e) {
             e.printStackTrace();
         } catch (SQLException e) {
             e.printStackTrace();
         }finally {
+            if (rs != null){
+                try {
+                    rs.close();
+                } catch (SQLException e) {
+                    e.printStackTrace();
+                }
+            }
             if (stmt != null){
                 try {
                     stmt.close();
@@ -209,17 +379,112 @@ public class OrgDsFileUser implements OrgDsGateway {
             }
 
         }
+        return l;
+
 
 
     }
 
-    public void utilDeleteOrgUnpublishedEvent(String org_username, String event_title){
+    public ArrayList<String> utilGetAllPastEvent(String par_username){
+        Statement stmt = null;
+        Connection conn = null;
+        ResultSet rs = null;
+        ArrayList l = new ArrayList<String>(0);
+        try {
+            Class.forName("com.mysql.jdbc.Driver");
+            conn = DriverManager.getConnection("jdbc:mysql://localhost:3306/db2", "root", "1234");
+            String sql = "select event_title from past_events_for_par where par_username = '" + par_username + "';";
+            stmt = conn.createStatement();
+            rs = stmt.executeQuery(sql);
+            System.out.println(sql);
+            while (rs.next()){
+                l.add(rs.getString(1));
+            }
+        } catch (ClassNotFoundException e) {
+            e.printStackTrace();
+        } catch (SQLException e) {
+            e.printStackTrace();
+        }finally {
+            if (rs != null){
+                try {
+                    rs.close();
+                } catch (SQLException e) {
+                    e.printStackTrace();
+                }
+            }
+            if (stmt != null){
+                try {
+                    stmt.close();
+                } catch (SQLException e) {
+                    e.printStackTrace();
+                }
+            }
+            if (conn != null){
+                try {
+                    conn.close();
+                } catch (SQLException e) {
+                    e.printStackTrace();
+                }
+            }
+
+        }
+        return l;
+
+    }
+    public ArrayList<String> utilGetAllUpcomingEvent(String par_username){
+        Statement stmt = null;
+        Connection conn = null;
+        ResultSet rs = null;
+        ArrayList l = new ArrayList<String>(0);
+        try {
+            Class.forName("com.mysql.jdbc.Driver");
+            conn = DriverManager.getConnection("jdbc:mysql://localhost:3306/db2", "root", "1234");
+            String sql = "select event_title from upcoming_events_for_par where par_username = '" + par_username + "';";
+            stmt = conn.createStatement();
+            rs = stmt.executeQuery(sql);
+            System.out.println(sql);
+            while (rs.next()){
+                l.add(rs.getString(1));
+            }
+        } catch (ClassNotFoundException e) {
+            e.printStackTrace();
+        } catch (SQLException e) {
+            e.printStackTrace();
+        }finally {
+            if (rs != null){
+                try {
+                    rs.close();
+                } catch (SQLException e) {
+                    e.printStackTrace();
+                }
+            }
+            if (stmt != null){
+                try {
+                    stmt.close();
+                } catch (SQLException e) {
+                    e.printStackTrace();
+                }
+            }
+            if (conn != null){
+                try {
+                    conn.close();
+                } catch (SQLException e) {
+                    e.printStackTrace();
+                }
+            }
+
+        }
+        return l;
+
+    }
+
+    public void utilPasswordUpdating(String par_username, String new_password){
         Statement stmt = null;
         Connection conn = null;
         try {
             Class.forName("com.mysql.jdbc.Driver");
             conn = DriverManager.getConnection("jdbc:mysql://localhost:3306/db2", "root", "1234");
-            String sql = "delete from unpublished_events_for_org where org_username = '" + org_username + "' and event_title = '" + event_title + "';";
+            String sql = "update parfile set password = '" + new_password + "' where username = '" + par_username + "';";
             stmt = conn.createStatement();
             int count = stmt.executeUpdate(sql);
             System.out.println(sql);
@@ -250,58 +515,14 @@ public class OrgDsFileUser implements OrgDsGateway {
             }
 
         }
-
     }
-
-    public void utilAddOrgUpcomingEvent(String org_username, String event_title){
+    public void utilNotificationUpdating(String par_username, String new_notification){
         Statement stmt = null;
         Connection conn = null;
         try {
             Class.forName("com.mysql.jdbc.Driver");
             conn = DriverManager.getConnection("jdbc:mysql://localhost:3306/db2", "root", "1234");
-            String sql = "insert into upcoming_events_for_org(org_username, event_title) values('" + org_username + "','" + event_title + "');" ;
-            stmt = conn.createStatement();
-            int count = stmt.executeUpdate(sql);
-            System.out.println(sql);
-            if (count > 0 ){
-                System.out.println("Success");
-            }
-            else {
-                System.out.println("Failure");
-            }
-
-        } catch (ClassNotFoundException e) {
-            e.printStackTrace();
-        } catch (SQLException e) {
-            e.printStackTrace();
-        }finally {
-            if (stmt != null){
-                try {
-                    stmt.close();
-                } catch (SQLException e) {
-                    e.printStackTrace();
-                }
-            }
-            if (conn != null){
-                try {
-                    conn.close();
-                } catch (SQLException e) {
-                    e.printStackTrace();
-                }
-            }
-
-        }
-
-
-    }
-
-    public void utilDeleteOrgUpcomingevent(String org_username, String event_title){
-        Statement stmt = null;
-        Connection conn = null;
-        try {
-            Class.forName("com.mysql.jdbc.Driver");
-            conn = DriverManager.getConnection("jdbc:mysql://localhost:3306/db2", "root", "1234");
-            String sql = "delete from upcoming_events_for_org where org_username = '" + org_username + "' and event_title = '" + event_title + "';";
+            String sql = "update parfile set notification = '" + new_notification + "' where username = '" + par_username + "';";
             stmt = conn.createStatement();
             int count = stmt.executeUpdate(sql);
             System.out.println(sql);
@@ -332,209 +553,32 @@ public class OrgDsFileUser implements OrgDsGateway {
             }
 
         }
-
     }
 
-    public ArrayList<String> utilGetAllFollowers(String org_username){
-        Statement stmt = null;
-        Connection conn = null;
-        ResultSet rs = null;
-        ArrayList l = new ArrayList<String>(0);
-        try {
-            Class.forName("com.mysql.jdbc.Driver");
-            conn = DriverManager.getConnection("jdbc:mysql://localhost:3306/db2", "root", "1234");
-            String sql = "select par_username from follow_org_par where org_username = '" + org_username + "';";
-            stmt = conn.createStatement();
-            rs = stmt.executeQuery(sql);
-            System.out.println(sql);
-//            rs.next();
-            while (rs.next()){
-                l.add(rs.getString(1));
-            }
-        } catch (ClassNotFoundException e) {
-            e.printStackTrace();
-        } catch (SQLException e) {
-            e.printStackTrace();
-        }finally {
-            if (rs != null){
-                try {
-                    rs.close();
-                } catch (SQLException e) {
-                    e.printStackTrace();
-                }
-            }
-            if (stmt != null){
-                try {
-                    stmt.close();
-                } catch (SQLException e) {
-                    e.printStackTrace();
-                }
-            }
-            if (conn != null){
-                try {
-                    conn.close();
-                } catch (SQLException e) {
-                    e.printStackTrace();
-                }
-            }
-
-        }
-        return l;
 
 
 
-    }
-    public ArrayList<String> utilGetUnpublishedEvents(String org_username){
-        Statement stmt = null;
-        Connection conn = null;
-        ResultSet rs = null;
-        ArrayList l = new ArrayList<String>(0);
-        try {
-            Class.forName("com.mysql.jdbc.Driver");
-            conn = DriverManager.getConnection("jdbc:mysql://localhost:3306/db2", "root", "1234");
-            String sql = "select event_title from unpublished_events_for_org where org_username = '" + org_username + "';";
-            stmt = conn.createStatement();
-            rs = stmt.executeQuery(sql);
-            System.out.println(sql);
-//            rs.next();
-            while (rs.next()){
-                l.add(rs.getString(1));
-            }
-        } catch (ClassNotFoundException e) {
-            e.printStackTrace();
-        } catch (SQLException e) {
-            e.printStackTrace();
-        }finally {
-            if (rs != null){
-                try {
-                    rs.close();
-                } catch (SQLException e) {
-                    e.printStackTrace();
-                }
-            }
-            if (stmt != null){
-                try {
-                    stmt.close();
-                } catch (SQLException e) {
-                    e.printStackTrace();
-                }
-            }
-            if (conn != null){
-                try {
-                    conn.close();
-                } catch (SQLException e) {
-                    e.printStackTrace();
-                }
-            }
-
-        }
-        return l;
 
 
 
-    }
-    public ArrayList<String> utilGetPastEvents(String org_username){
-        Statement stmt = null;
-        Connection conn = null;
-        ResultSet rs = null;
-        ArrayList l = new ArrayList<String>(0);
-        try {
-            Class.forName("com.mysql.jdbc.Driver");
-            conn = DriverManager.getConnection("jdbc:mysql://localhost:3306/db2", "root", "1234");
-            String sql = "select event_title from past_events_for_org where org_username = '" + org_username + "';";
-            stmt = conn.createStatement();
-            rs = stmt.executeQuery(sql);
-            System.out.println(sql);
-//            rs.next();
-            while (rs.next()){
-                l.add(rs.getString(1));
-            }
-        } catch (ClassNotFoundException e) {
-            e.printStackTrace();
-        } catch (SQLException e) {
-            e.printStackTrace();
-        }finally {
-            if (rs != null){
-                try {
-                    rs.close();
-                } catch (SQLException e) {
-                    e.printStackTrace();
-                }
-            }
-            if (stmt != null){
-                try {
-                    stmt.close();
-                } catch (SQLException e) {
-                    e.printStackTrace();
-                }
-            }
-            if (conn != null){
-                try {
-                    conn.close();
-                } catch (SQLException e) {
-                    e.printStackTrace();
-                }
-            }
-
-        }
-        return l;
 
 
 
-    }
-    public ArrayList<String> utilGetUpcomingEvents(String org_username){
-        Statement stmt = null;
-        Connection conn = null;
-        ResultSet rs = null;
-        ArrayList l = new ArrayList<String>(0);
-        try {
-            Class.forName("com.mysql.jdbc.Driver");
-            conn = DriverManager.getConnection("jdbc:mysql://localhost:3306/db2", "root", "1234");
-            String sql = "select event_title from upcoming_events_for_org where org_username = '" + org_username + "';";
-            stmt = conn.createStatement();
-            rs = stmt.executeQuery(sql);
-            System.out.println(sql);
-//            rs.next();
-            while (rs.next()){
-                l.add(rs.getString(1));
-            }
-        } catch (ClassNotFoundException e) {
-            e.printStackTrace();
-        } catch (SQLException e) {
-            e.printStackTrace();
-        }finally {
-            if (rs != null){
-                try {
-                    rs.close();
-                } catch (SQLException e) {
-                    e.printStackTrace();
-                }
-            }
-            if (stmt != null){
-                try {
-                    stmt.close();
-                } catch (SQLException e) {
-                    e.printStackTrace();
-                }
-            }
-            if (conn != null){
-                try {
-                    conn.close();
-                } catch (SQLException e) {
-                    e.printStackTrace();
-                }
-            }
-
-        }
-        return l;
 
 
 
-    }
 
-    public String utilGetPassword(String username) {
-        //Return the password of the entered organizer user
-        //Used for login password check
+
+
+
+
+
+
+
+
+
+
+    public String getPassword(String username){
         Statement stmt = null;
         Connection conn = null;
         ResultSet rs = null;
@@ -542,8 +586,9 @@ public class OrgDsFileUser implements OrgDsGateway {
         try {
             Class.forName("com.mysql.jdbc.Driver");
             conn = DriverManager.getConnection("jdbc:mysql://localhost:3306/db2", "root", "1234");
+            String sql = "select password from parfile where username = '" + username + "';";
             stmt = conn.createStatement();
-            rs = stmt.executeQuery("select password from orgfile where username = \"" + username + "\";");
+            rs = stmt.executeQuery(sql);
             rs.next();
             password = rs.getString("password");
         } catch (ClassNotFoundException e) {
@@ -552,22 +597,22 @@ public class OrgDsFileUser implements OrgDsGateway {
         } catch (SQLException e) {
             System.out.println("SQL");
             e.printStackTrace();
-        } finally {
-            if (rs != null) {
+        }finally {
+            if(rs != null){
                 try {
                     rs.close();
                 } catch (SQLException e) {
                     e.printStackTrace();
                 }
             }
-            if (stmt != null) {
+            if(stmt != null){
                 try {
                     stmt.close();
                 } catch (SQLException e) {
                     e.printStackTrace();
                 }
             }
-            if (conn != null) {
+            if(conn != null){
                 try {
                     conn.close();
                 } catch (SQLException e) {
@@ -577,83 +622,41 @@ public class OrgDsFileUser implements OrgDsGateway {
         }
         return password;
     }
-
-    public void utilPasswordUpdating(String org_username, String new_password){
-        Statement stmt = null;
-        Connection conn = null;
-        try {
-            Class.forName("com.mysql.jdbc.Driver");
-            conn = DriverManager.getConnection("jdbc:mysql://localhost:3306/db2", "root", "1234");
-            String sql = "update orgfile set password = '" + new_password + "' where username = '" + org_username + "';";
-            stmt = conn.createStatement();
-            int count = stmt.executeUpdate(sql);
-            System.out.println(sql);
-            if (count > 0) {
-                System.out.println("Success");
-            } else {
-                System.out.println("Failure");
-            }
-
-        } catch (ClassNotFoundException e) {
-            e.printStackTrace();
-        } catch (SQLException e) {
-            e.printStackTrace();
-        } finally {
-            if (stmt != null) {
-                try {
-                    stmt.close();
-                } catch (SQLException e) {
-                    e.printStackTrace();
-                }
-            }
-            if (conn != null) {
-                try {
-                    conn.close();
-                } catch (SQLException e) {
-                    e.printStackTrace();
-                }
-            }
-
-        }
-    }
-
-    public ArrayList<String> utilOrganizerSearch(String about_name){
-        //This is method is used for searching method of the website
+    public String getNotification(String username){
         Statement stmt = null;
         Connection conn = null;
         ResultSet rs = null;
-        ArrayList<String> l = new ArrayList<String>(0);
+        String notification = null;
         try {
-
             Class.forName("com.mysql.jdbc.Driver");
             conn = DriverManager.getConnection("jdbc:mysql://localhost:3306/db2", "root", "1234");
+            String sql = "select notification from parfile where username = '" + username + "';";
             stmt = conn.createStatement();
-            rs = stmt.executeQuery("select username from orgfile where username like \"%" + about_name + "%\";");
-            while (rs.next()) {
-                l.add(rs.getString("username"));
-            }
+            rs = stmt.executeQuery(sql);
+            rs.next();
+            notification = rs.getString("notification");
         } catch (ClassNotFoundException e) {
             System.out.println("NotFound");
             e.printStackTrace();
         } catch (SQLException e) {
             System.out.println("SQL");
             e.printStackTrace();
-        } finally {
-            if (rs != null) {
+        }finally {
+            if(rs != null){
                 try {
                     rs.close();
                 } catch (SQLException e) {
                     e.printStackTrace();
                 }
             }
-            if (stmt != null) {
+            if(stmt != null){
                 try {
                     stmt.close();
                 } catch (SQLException e) {
                     e.printStackTrace();
                 }
             }
-            if (conn != null) {
+            if(conn != null){
                 try {
                     conn.close();
                 } catch (SQLException e) {
@@ -661,85 +664,41 @@ public class OrgDsFileUser implements OrgDsGateway {
                 }
             }
         }
-        return l;
-    }
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-    public ArrayList<String> organizerSearch(String about_name){
-        return utilOrganizerSearch(about_name);
-    }
-
-
-    public String getPassword(String username) {
-        return utilGetPassword(username);
-    }
-
-    public void setPassword(String username, String new_password){
-        //
-        utilPasswordUpdating(username, new_password);
-    }
-
-
-    public ArrayList<String> getUnpublishedEvents(String username){
-        return utilGetUnpublishedEvents(username);
-    }
-
-    public ArrayList<String> getPastEvent(String username){
-        return utilGetPastEvents(username);
+        return notification;
     }
 
     public ArrayList<String> getUpcomingEvents(String username){
-        return utilGetUpcomingEvents(username);
+        return utilGetAllUpcomingEvent(username);
     }
 
-    public ArrayList<String> getFollowers(String username){
-        return utilGetAllFollowers(username);
+    public ArrayList<String> getPastEvents(String username){
+        return utilGetAllPastEvent(username);
     }
 
-    public void createAnEvent(String org_username,
-                              String title,
-                              int status,
-                              int event_type,
-                              String description,
-                              String location,
-                              String image_path,
-                              int year,
-                              int month,
-                              int day,
-                              int hour,
-                              int minute){
-        EventFileUser temp_eventfileuser = new EventFileUser();
-        temp_eventfileuser.utilStoreEvent(title, status, event_type, description, location, image_path, year, month, day, hour, minute);
-        if (status == 0){
-            //Unpublished
-            utilAddOrgUnpublishedEvent(org_username,title);
-        }
-        else if(status == 1){
-            //Past
-            utilAddOrgPastEvent(org_username,title);
-        }
-        else{
-            //Upcoming
-            utilAddOrgUpcomingEvent(org_username,title);
+    public ArrayList<String> getFollowedOrg(String username){
+        return utilGetAllFollowing(username);
+    }
 
-        }
+    public void setPassword(String username, String new_password){
+        utilPasswordUpdating(username, new_password);
     }
-    public void deleteAnEvent(String username, String title){
-        EventFileUser temp_eventfileuser = new EventFileUser();
-        temp_eventfileuser.deleteEvent(title);
+
+    public void setNofication(String username, String new_notification){
+        utilNotificationUpdating(username,new_notification);
     }
+
+    public void followOrg(String par_username, String org_username){
+        utiladdparfollowing(par_username,org_username);
+    }
+
+    public void registerEvent(String par_username, String title){
+        utilAddParUpcomingEvent(par_username,title);
+    }
+
+    public void leaveEvent(String par_username, String title){
+        utilDeleteParUpcomingevent(par_username,title);
+    }
+
 
     public boolean checkIfUsernameExist(String username){
         Statement stmt = null;
@@ -749,7 +708,7 @@ public class OrgDsFileUser implements OrgDsGateway {
         try {
             Class.forName("com.mysql.jdbc.Driver");
             conn = DriverManager.getConnection("jdbc:mysql://localhost:3306/db2", "root", "1234");
-            String sql = "select exists(select * from orgfile where username = '" + username + "');";
+            String sql = "select exists(select * from parfile where username = '" + username + "');";
             stmt = conn.createStatement();
             rs = stmt.executeQuery(sql);
             rs.next();
@@ -788,33 +747,33 @@ public class OrgDsFileUser implements OrgDsGateway {
         return WhetherExist;
     }
 
-    public void createOrg(String username, String password){
-        utilStoreOrg(username,password);
+    public void createPar(String username, String password){
+        utilStorePar(username, password);
     }
-    public void deleteOrg(String username){
+
+    public void deletePar(String username){
         //First delete relationships with events
-        //Then delete relationships with participants
+        //Then delete relationships with organizers
         //Then delete itself
-        ParFileUser temp_parfileuser = new ParFileUser();
-        ArrayList<String> All_Unpublished = utilGetUnpublishedEvents(username);
-        for (int i=0; i<All_Unpublished.size();i++){
-            utilDeleteOrgUnpublishedEvent(username, All_Unpublished.get(i));
+
+        //First breaking relationship with past events
+        ArrayList<String> All_past_events = utilGetAllPastEvent(username);
+        for (int i=0;i<All_past_events.size();i++){
+            utilDeleteParPastevent(username,All_past_events.get(i));
         }
-        ArrayList<String> All_Past = utilGetPastEvents(username);
-        for (int i=0; i<All_Past.size();i++){
-            utilDeleteOrgPastEvent(username, All_Past.get(i));
-        }
-        ArrayList<String> All_upcoming = utilGetUpcomingEvents(username);
-        for (int i=0; i<All_Past.size();i++){
-            utilDeleteOrgUpcomingevent(username, All_upcoming.get(i));
+        ArrayList<String> All_upcoming_events = utilGetAllUpcomingEvent(username);
+        for (int i=0;i<All_upcoming_events.size();i++){
+            utilDeleteParUpcomingevent(username, All_upcoming_events.get(i));
         }
 
-        ArrayList<String> All_followers = utilGetAllFollowers(username);
-        for (int i=0; i<All_followers.size();i++){
-            temp_parfileuser.utilDeleteParFollowOrg(All_followers.get(i), username);
+        //Then break relationship with followings
+        ArrayList<String> All_following = utilGetAllFollowing(username);
+        for (int i=0;i<All_following.size();i++){
+            utilDeleteParFollowOrg(username, All_following.get(i));
         }
 
-        utilDeleteOrg(username);
+        utilDeletePar(username);
+
     }
 
 }
